+++ src/page/Homepage/Homepage.js
@@ -5,15 +5,22 @@ import Featured from "../../components/featured/Featured";
 import Footer from "../../components/footer/Footer";
 import Header1 from "../../components/Header/Header1";
 import Navbar from "../../components/navbar/Navbar";
+import API from "../../utils/API";
 import "./Homepage.scss";
 
 class Homepage extends React.Component {
   constructor(props) {
     super();
+    this.state = { produk: null };
   }
 
   componentDidMount = () => {
     window.scrollTo(0, 0);
+    API.get("/produk?limit=20").then((res) => {
+      if (res.status == 200) {
+        this.setState({ produk: res.data });
+      }
+    });
   };
 
   render() {
@@ -29,7 +36,7 @@ class Homepage extends React.Component {
         </div>
 
         <div className="container">
-          <Catalog />
+          {this.state.produk ? <Catalog data={this.state.produk} /> : ""}
 
           <div className="text-center">
             <Link to="/catalog" className="btn btn-secondary mb-4">